Migrated to newer SDL Library version and Updated Build System
- Migrated codebase from SDL2 to SDL3
    - Removed SDL_image initialization and deinitialization code
    - Updated all #include preprocessor directives for SDL3, SDL3_image, and SDL3_ttf
    - Added LOG_PRIORITY_INVALID and LOG_PRIORITY_TRACE to the log priority enum
    - Replaced all instances of 'SDL_Rect' with 'SDL_FRect'
    - Updated the 'Audio Manager' class for the SDL3 audio API
    - Updated the Renderer class for the SDL3 rendering API
    - Updated SDL event handling
    - Replaced the SDL2 backends for ImGui with the SDL3 backends
    - Updated 'CMakeLists.txt' to fetch SDL3 libraries

- Integrated a declarative manifest into the build system
    - Added 'vcpkg.json' maniest to declare the project dependencies and versions
    - Removed explicit vcpkg install commands from 'build.bat'
    - Replaced manual build of Box2D with 'find_package' in 'CMakeLists.txt'

+++ sources/Media/Button.hpp
@@ -22,10 +22,10 @@ public:
     Button() = default;
 
     // Determines if the mouse is currently hovering over the button
-    bool IsActive(uint16_t mouseX, uint16_t mouseY) const;
+    bool IsActive(float mouseX, float mouseY) const;
 
     // Determines if the button has been clicked
-    bool IsClicked(uint16_t mouseX, uint16_t mouseY, const SDL_Event& event) const;
+    bool IsClicked(float mouseX, float mouseY, const SDL_Event& event) const;
 
 public:
     // The name of the button
